+++ pnet-data-api-java/src/main/java/pnet/data/api/company/CompanyDataDTO.java
@@ -24,6 +24,7 @@ import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import pnet.data.api.GeoPoint;
 import pnet.data.api.companygroup.CompanyGroupMemberLinkDTO;
+import pnet.data.api.util.PnetDataApiUtils;
 import pnet.data.api.util.WithCompanyId;
 import pnet.data.api.util.WithMatchcode;
 import pnet.data.api.util.WithTenants;
@@ -49,13 +50,20 @@ public class CompanyDataDTO implements WithCompanyId, WithMatchcode, WithTenants
     @ApiModelProperty(notes = "The tenant (Portal-ID), in which this company gets administrated.")
     private String administrativeTenant;
 
-    @ApiModelProperty(notes = "The name of the company.")
+    @ApiModelProperty(
+        notes = "The label of the company (either the marketing name or a combination of name and affix).")
+    private String label;
+
+    @ApiModelProperty(notes = "The name of the company. Deprecated: use label instead.")
+    @Deprecated
     private String name;
 
-    @ApiModelProperty(notes = "The name affix of the company.")
+    @ApiModelProperty(notes = "The name affix of the company. Deprecated: use label instead.")
+    @Deprecated
     private String nameAffix;
 
-    @ApiModelProperty(notes = "The marketing name of the company.")
+    @ApiModelProperty(notes = "The marketing name of the company. Deprecated: use label instead.")
+    @Deprecated
     private String marketingName;
 
     @ApiModelProperty(notes = "Groups this company is part of.")
@@ -212,31 +220,64 @@ public class CompanyDataDTO implements WithCompanyId, WithMatchcode, WithTenants
         this.administrativeTenant = administrativeTenant;
     }
 
+    public String getLabel()
+    {
+        return label;
+    }
+
+    public String getLabelWithNumber()
+    {
+        return PnetDataApiUtils.toCompanyLabelWithNumber(companyNumber, label);
+    }
+
+    public void setLabel(String label)
+    {
+        this.label = label;
+    }
+
+    /**
+     * @return the name of the company
+     * @deprecated use the label instead
+     */
+    @Deprecated
     public String getName()
     {
         return name;
     }
 
+    @Deprecated
     public void setName(String name)
     {
         this.name = name;
     }
 
+    /**
+     * @return the name affix of the company
+     * @deprecated use the label instead
+     */
+    @Deprecated
     public String getNameAffix()
     {
         return nameAffix;
     }
 
+    @Deprecated
     public void setNameAffix(String nameAffix)
     {
         this.nameAffix = nameAffix;
     }
 
+    /**
+     * @return the marketing name of the company
+     * @deprecated use the label instead
+     */
+    @Deprecated
     public String getMarketingName()
     {
         return marketingName;
     }
 
+    @Deprecated
     public void setMarketingName(String marketingName)
     {
         this.marketingName = marketingName;