+++ pnet-data-api-java/src/main/java/pnet/data/api/company/CompanyItemDTO.java
@@ -23,6 +23,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import pnet.data.api.GeoPoint;
+import pnet.data.api.util.PnetDataApiUtils;
 import pnet.data.api.util.WithCompanyId;
 import pnet.data.api.util.WithLastUpdate;
 import pnet.data.api.util.WithMatchcode;
@@ -51,13 +52,20 @@ public class CompanyItemDTO
     @ApiModelProperty(notes = "The tenant (Portal-ID), in which this company gets administrated.")
     private final String administrativeTenant;
 
-    @ApiModelProperty(notes = "The name of the company.")
+    @ApiModelProperty(
+        notes = "The label of the company (either the marketing name or a combination of name and affix).")
+    private final String label;
+
+    @ApiModelProperty(notes = "The name of the company. Deprecated: use label instead.")
+    @Deprecated
     private final String name;
 
-    @ApiModelProperty(notes = "The name affix of the company.")
+    @ApiModelProperty(notes = "The name affix of the company. Deprecated: use label instead.")
+    @Deprecated
     private final String nameAffix;
 
-    @ApiModelProperty(notes = "The marketing name of the company.")
+    @ApiModelProperty(notes = "The marketing name of the company. Deprecated: use label instead.")
+    @Deprecated
     private final String marketingName;
 
     @ApiModelProperty(notes = "Valid tenants of the company (also known as Portal-ID).")
@@ -103,9 +111,9 @@ public class CompanyItemDTO
     private final double score;
 
     public CompanyItemDTO(@JsonProperty("companyId") Integer companyId, @JsonProperty("matchcode") String matchcode,
-        @JsonProperty("administrativeTenant") String administrativeTenant, @JsonProperty("name") String name,
-        @JsonProperty("nameAffix") String nameAffix, @JsonProperty("marketingName") String marketingName,
-        @JsonProperty("tenants") Collection<String> tenants,
+        @JsonProperty("administrativeTenant") String administrativeTenant, @JsonProperty("label") String label,
+        @JsonProperty("name") String name, @JsonProperty("nameAffix") String nameAffix,
+        @JsonProperty("marketingName") String marketingName, @JsonProperty("tenants") Collection<String> tenants,
         @JsonProperty("brands") Collection<CompanyBrandLinkDTO> brands,
         @JsonProperty("companyNumber") String companyNumber, @JsonProperty("street") String street,
         @JsonProperty("city") String city, @JsonProperty("postalCode") String postalCode,
@@ -118,6 +126,7 @@ public class CompanyItemDTO
         this.companyId = companyId;
         this.matchcode = matchcode;
         this.administrativeTenant = administrativeTenant;
+        this.label = label;
         this.name = name;
         this.nameAffix = nameAffix;
         this.marketingName = marketingName;
@@ -159,16 +168,41 @@ public class CompanyItemDTO
         return administrativeTenant;
     }
 
+    public String getLabel()
+    {
+        return label;
+    }
+
+    public String getLabelWithNumber()
+    {
+        return PnetDataApiUtils.toCompanyLabelWithNumber(companyNumber, label);
+    }
+
+    /**
+     * @return the name of the company
+     * @deprecated use the label instead
+     */
+    @Deprecated
     public String getName()
     {
         return name;
     }
 
+    /**
+     * @return the name affix of the company
+     * @deprecated use the label instead
+     */
+    @Deprecated
     public String getNameAffix()
     {
         return nameAffix;
     }
 
+    /**
+     * @return the marketing name of the company
+     * @deprecated use the label instead
+     */
+    @Deprecated
     public String getMarketingName()
     {
         return marketingName;